+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnPlaylistQuery.kt
@@ -81,6 +81,11 @@ class OnPlaylistQuery : ViewModel() {
                 for (playlistMedia in cursor.columnNames) {
                     playlistData[playlistMedia] = helper.loadPlaylistItem(playlistMedia, cursor)
                 }
+
+                // Count and add the number of songs for every playlist.
+                val mediaCount = helper.getMediaCount(1, playlistData["_id"].toString(), resolver)
+                playlistData["num_of_songs"] = mediaCount
+
                 playlistList.add(playlistData)
             }