+++ on_audio_query_platform_interface/lib/details/models/genre_model.dart
@@ -13,6 +13,9 @@ class GenreModel {
   /// Return [genre] name
   String get genre => _info["name"];
 
+  ///Return genre [numOfSongs]
+  String get numOfSongs => _info["num_of_songs"];
+
   /// Return a map with all [keys] and [values] from specific genre.
   Map get getMap => _info;